global: use fixtures to populate the db
* Populates the database using pytest fixtures. All the fixtures are
  stored in JSON files inside "tests/myclaimstore/". (addresses #20)

* Redefines the commands `initdb` (creates the db tables and
  populates it with basic data (predicates, persistent identifiers
  and claimants) from "tests/myclaimstore/config/") and `populatedb`
  (which populates the db with claims from "tests/myclaimstore/data").

+++ claimstore/ext/sqlalchemy.py
@@ -37,21 +37,16 @@ models = RegistryProxy(
 def setup_app(app):
     """Setup sqlalchemy."""
     # Add extension CLI to application.
-    app.cli.add_command(initdb)
     app.cli.add_command(dropalldb)
     db.init_app(app)
 
 
 @click.command()
 @with_appcontext
-def initdb():
-    """Initialise database."""
-    list(models)  # load all modules
-    db.create_all()
-
-
-@click.command()
-@with_appcontext
 def dropalldb():
     """Drop database."""
-    db.drop_all()
+    if click.confirm('Are you sure you want to drop the whole database?'):
+        db.drop_all()
+        click.echo('Database dropped')
+    else:
+        click.echo('Command aborted')
